Making a To-Do List
I'm really happy with how this turned out! 
I made changes in the code I found online and 
created a to-do list that I can add items to,
complete items with a strike-through line, 
and delete items. 
I did not get the final steps of saving to local storage 
and then fetching them back.